Checkpoint 1- Complete
I didn't see a finished version of the produceAnswer method, so I added what I did (I did it just so I knew I understood how to do it). the main method is updated to work for checkpoint 2 because I thought checkpoint 2, not 1, was due tomorrow-- oops. I tested the code and everything worked, so if you have any questions about it, please feel free to ask. Of course, if you also have a working method, we can turn that in instead- I just added this in case we didn't have time tomorrow to really put our code together.