Finished multiplication
I finished the multiplication and added two methods. One converts improper fractions to mixed numbers all in one, and the other finds the greatest common factor of two numbers. This is for simplifying the two numbers. It didn't ask to convert the improper fractions to mixed numbers, but I just thought I would add that in there.