Updated Error Catches
I made the error catches more clear and specific. It still errors when there's a space before or after the expression, but I stated in the error catch that you shouldn't do that. Also, the error catch does not catch multiple operators (i.e. 1 ++ 2), unfortunately. We would have to recode the method calls in order to fix that.